sns-lambda-cdk/src/requirements.txt
aws-cdk.core
aws-cdk.aws-lambda
aws-cdk.aws-logs
aws-cdk.aws-iam
aws-cdk.aws-sns
aws-cdk.aws-sns-subscriptions